+++ src/VulnAuction.sol
@@ -60,7 +60,7 @@ contract VulnAuction is ISimpleAuction {
         uint256 bestAmount;
         bool anyReveal = false;
 
-        for (uint i = 0; i < candidates.length; i++) {
+        for (uint256 i = 0; i < candidates.length; i++) {
             address bidder = candidates[i];
             Reveal storage r = reveals[auctionId][bidder];
             if (!r.revealed) continue;
